rename package name example.demo -> wafflytime
