+++ build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 	kotlin("plugin.jpa") version "1.7.21"
 }
 
-group = "com.example"
+group = "com.wafflytime"
 version = "0.0.1-SNAPSHOT"
 java.sourceCompatibility = JavaVersion.VERSION_17
