+++ src/main/kotlin/com/wafflytime/dto/VerifyEmailCode.kt
@@ -1,3 +1,3 @@
-package com.example.demo.dto
+package com.wafflytime.dto
 
 data class VerifyEmailCode(val code: String)
