+++ src/main/kotlin/com/wafflytime/exception/WafflyTimeExceptionHandler.kt
@@ -1,4 +1,4 @@
-package com.example.demo.exception
+package com.wafflytime.exception
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
