+++ src/test/kotlin/com/wafflytime/DemoApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.example.demo
+package com.wafflytime
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
